# Speed up _string_helpers init.