@rossbar 2023-05-18
Are we sure that importing string is still slow? I see the comment above this section but it's at least 5 years old. It'd be good to check!