@ttsugriy 2023-05-18
I'm not sure how to check this since it likely heavily depends on the used interpreter.