@ngoldbaum 2023-05-22
It's still a lot slower to do the import on Python 3.11:

```
$ pyperf timeit "import string"
.....................
Mean +- std dev: 95.7 ns +- 1.4 ns
```

```
$ pyperf timeit "_all_chars = tuple(map(chr, range(256)))
_ascii_upper = _all_chars[65:65+26]
_ascii_lower = _all_chars[97:97+26]
LOWER_TABLE = \"\".join(_all_chars[:65] + _ascii_lower + _all_chars[65+26:])
UPPER_TABLE = \"\".join(_all_chars[:97] + _ascii_upper + _all_chars[97+26:])"
.....................

Mean +- std dev: 13.1 us +- 3.7 us
```