@rossbar 2023-05-22
> It's still a lot slower to do the import on Python 3.11:

Technically the `pyperf` things above show the opposite: 95 ns vs. 13 us. But anyways that's a bit misleading because `pyperf` is probably reporting results from imports after caching.