@ngoldbaum 2023-05-22
> Technically the pyperf things above show the opposite: 95 ns vs. 13 us.

Argh, sorry for the lack of care, I wasn't reading very carefully. You're right about the caching too, pyperf has a warmup run.

Profiling with `-x importtime` says that `import string` takes about 800 us on my laptop.